% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparePWM.R
\name{comparePWM}
\alias{comparePWM}
\title{Compare PWM matrices}
\usage{
comparePWM(name, pre1 = "Mmusculus-JASPAR_CORE-",
  pre2 = "Mmusculus-JASPAR_2014-", db = tfMotifs)
}
\arguments{
\item{name}{the name of the motif without the dataSource specific prefix}

\item{pre1}{the prefix from the first dataSource}

\item{pre2}{the prefix from the second dataSource}

\item{db}{the common database which contains PWM information from the two dataSources.
Should be in \code{MotifDb} format.}
}
\value{
A single row \code{data_frame} with the supplied \code{name}
and the \code{logical} value in the variable \code{identical}
}
\description{
Compare PWM matrices
}
\details{
Designed to compare two matrices with an identical name from two databases,
and to check if they are the same
}
\author{
Steve Pederson
}

