% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparePWMFreqs.R
\name{comparePWMFreqs}
\alias{comparePWMFreqs}
\title{Compare the matches to a set of PWMs in two sets of sequences}
\usage{
comparePWMFreqs(refSeq, testSeq, motifs, minScore = "90\%",
  testCounts = NULL, useFisher = FALSE)
}
\arguments{
\item{refSeq}{the set of reference sequences as an \code{\link{XStringSet}}}

\item{testSeq}{the set of test sequences as an \code{\link{XStringSet}}}

\item{motifs}{the set of motifs as a \code{\link{MotifList}}}

\item{minScore}{the minimum matching score for the PWM}

\item{testCounts}{if resampling over multiple reference sequences, supply this as
the output from \code{\link{countViews}}}

\item{useFisher}{\code{logical}. Adds the p-values from the \code{\link{fisher.test}} function to the output}
}
\value{
Returns a \code{tbl_df} with columns \code{motif}, \code{testFreq}, \code{refFreq} & \code{oddsRatio},
with the additional column \code{p} if \code{useFisher=TRUE}.
If the argument \code{useFisher=TRUE}, the odds ratio will be the conditional MLE estimate as calculated using that function,
whereas otherwise it will be the sample odds ratio.
}
\description{
Compare the frequency of matches to a set of motifs within two sets of sequences
}
\details{
Used to compare matches to the same set of motifs within two sets of sequences.

If using to bootstrap multiple sets of reference sequences, set \code{useFisher = FALSE} and
 use \code{\link{lapply}} (or it's relatives) to specify multiple sets of \code{XStringSet}s.

If simply testing two sets of sequences, set \code{iseFisher = TRUE} to add p-values
 from Fisher tests to every motif
}
\author{
Steve Pederson
}
\seealso{
\code{\link{tbl_df}}
}

