% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.percent.R
\name{is.percent}
\alias{is.percent}
\title{Check percentage format}
\usage{
is.percent(x, min = -Inf, max = Inf)
}
\arguments{
\item{x}{the object to test}

\item{min}{the minimum valid percentage as an integer}

\item{max}{the maximum valid percentage as an integer}
}
\value{
A \code{logical} vector
}
\description{
Check the percentage format as obtained by the function \code{percent} from the \code{scales} package
}
\details{
Checks that a value is specified as a percentage, as obtained by the function \code{\link{percent}}
}
\examples{
is.percent("30\%")
is.percent("200\%")
is.percent("200\%", max=100)
is.percent(100)
}
\author{
Steve Pederson
}

