% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quickFuns.R
\name{quickOR}
\alias{quickFisher}
\alias{quickOR}
\title{Quick calculations used in the package}
\usage{
quickOR(x)

quickFisher(x)
}
\arguments{
\item{x}{a 2x2 matrix}
}
\description{
Quick calculations of the unconditional odds ratio and fisher.test for 2x2 matrices only
}
\details{
Quick calculations designed to be utilised in the function \code{\link{comparePWMFreqs}}
The function \code{quickFisher} simply formats the output for consistency within this function
}
\author{
Steve Pederson
}

